+++ icegraph/renderer/__init__.py
@@ -3,12 +3,4 @@
 
 from .plots import CDFPlot, PDFPlot, ChargeDistributionPlot, ParityPlot, BiasPlot, ConfusionMatrixPlot, ROCPlot
 
-CDFPlot.__module__ = __name__
-PDFPlot.__module__ = __name__
-ChargeDistributionPlot.__module__ = __name__
-ParityPlot.__module__ = __name__
-BiasPlot.__module__ = __name__
-ConfusionMatrixPlot.__module__ = __name__
-ROCPlot.__module__ = __name__
-
 __all__ = ["CDFPlot", "PDFPlot", "ChargeDistributionPlot", "ParityPlot", "BiasPlot", "ConfusionMatrixPlot", "ROCPlot"]