+++ icegraph/types/__init__.py
@@ -2,7 +2,6 @@
 # Developed by Taylor St Jean
 
 from .metrics import ComputedMetrics
+from .callbacks import MetricsPlotMethod
 
-ComputedMetrics.__module__ = __name__
-
-__all__ = ["ComputedMetrics"]
+__all__ = ["ComputedMetrics", "MetricsPlotMethod"]