+++ icegraph/types/callbacks.py
@@ -1,8 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .tasks import resolve_strategy
+from typing import Callable
 
-resolve_strategy.__module__ = __name__
 
-__all__ = ["resolve_strategy"]
+MetricsPlotMethod = Callable[..., None]